Move edit command classes into logic.commands.edit package
