+++ src/main/java/seedu/address/logic/commands/edit/EditLectureCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.edit;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_LECTURE_DOES_NOT_EXIST;
@@ -9,6 +9,8 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
+import seedu.address.commons.util.CollectionUtil;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.lecture.Lecture;
@@ -101,6 +103,15 @@ public class EditLectureCommand extends EditCommand {
          */
         public EditLectureDescriptor(EditLectureDescriptor toCopy) {
             setName(toCopy.name);
+        }
+
+        /**
+         * Returns true if at least one field is edited.
+         *
+         * @return True if at least one field is edited. Otherwise, false.
+         */
+        public boolean isAnyFieldEdited() {
+            return CollectionUtil.isAnyNonNull(name);
         }
 
         public Optional<LectureName> getName() {
