+++ src/main/java/seedu/address/logic/parser/EditCommandParser.java
@@ -7,13 +7,13 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_LECTURE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_MODULE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 
-import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.EditLectureCommand;
-import seedu.address.logic.commands.EditLectureCommand.EditLectureDescriptor;
-import seedu.address.logic.commands.EditModuleCommand;
-import seedu.address.logic.commands.EditModuleCommand.EditModuleDescriptor;
-import seedu.address.logic.commands.EditVideoCommand;
-import seedu.address.logic.commands.EditVideoCommand.EditVideoDescriptor;
+import seedu.address.logic.commands.edit.EditCommand;
+import seedu.address.logic.commands.edit.EditLectureCommand;
+import seedu.address.logic.commands.edit.EditModuleCommand;
+import seedu.address.logic.commands.edit.EditVideoCommand;
+import seedu.address.logic.commands.edit.EditLectureCommand.EditLectureDescriptor;
+import seedu.address.logic.commands.edit.EditModuleCommand.EditModuleDescriptor;
+import seedu.address.logic.commands.edit.EditVideoCommand.EditVideoDescriptor;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.lecture.LectureName;
 import seedu.address.model.module.ModuleCode;
